♻️ Give width and height one type across the graphics namespace
The namespace measured the same thing four ways: Canvas.Width was int, Mode
was uint, Image was uint, Font is byte. Consumers cast both directions inside
one expression, DevKernel's overlay wrote `(uint)(SquareSize + Margin * 2)` to
compare against a resolution, and the ring's own code carried 15 `(int)Mode.
Width` casts.

That split was not only noise. DrawImage and DrawImageAlpha computed their
clip as `(uint)(Mode.Width - x)`: uint minus int is evaluated in long, and the
cast back is unchecked, so an image starting past the right edge got a bound
of about four billion instead of a negative one. That is fixed here by the
types rather than by the cast the previous commit added.

Mode and Image are int now, matching Canvas and System.Drawing, where every
pixel dimension is signed for exactly this reason: they are constantly mixed
with signed coordinates. The change removed 12 casts and added 3, all at real
uint boundaries in the SVGA II and GOP drivers.

Font keeps byte. It widens to int implicitly, so it forces no cast and cannot
wrap, and 255 is a real bound on a bitmap glyph.

+++ examples/DevKernel/Diagnostics/SystemMonitor.cs
@@ -38,7 +38,7 @@ internal static class SystemMonitor
 
         if (KernelFeatures.Mouse)
         {
-            MouseManager.SetScreenSize((int)canvas.Mode.Width, (int)canvas.Mode.Height);
+            MouseManager.SetScreenSize(canvas.Mode.Width, canvas.Mode.Height);
         }
 
         int x = OverlayLayout.TextMarginPx;
